Point to tags for archival purposes.
These tags capture the current state (as of today) of the branches they are pointing to. It will be an exercise left to the user to dig out specific hashes valid on the date of the last commit since these changes were tracked together with the top of a branch instead of a specific hash during this phase of RRFS development.